Read light DOM configuration in attach instead of ready
If you create a chart using Javascript, then the light DOM
is not be setup when the ready is fired, and thus the chart
will always be empty.

This change modifies the behavior so the light DOM is ready when
the <vaadin-some-chart> is attached to the document. At this point
the light DOM must be ready.

Inner <data-series> cannot use the attached event as they are never attached
to the document, they are only added to the light DOM.