Prevent caret from moving on arrow up and down.
Was not able to add a reliable test for this, seems Edge is the only one
browser that actually moves the caret both on input field changes and arrow
events. IE11 doesn't move the caret at all and the remaining browsers ignore
caret movements on generated arrow up/down events.