Don't scroll to selected item when items are set
Calling scrollToIndex on iron-list while it is rendering the list causes
it sometimes to scroll in a totally wrong position, especially if the list
size changes. Scrolling is called because we try to focus the selected item
when items are set.

To fix the issue, instead of calling _scrollIntoView asyncronously, we can
actually remove the focusing when items are set. It is enough to set the focus
only when there is an exact match when filtering or when a value is set.