Added dependency: iron-overlay-behavior^1.3.2
In the iron-overlay-behavior 1.3.2, there's a change that adds
passing the original event in `'iron-overlay-canceled'` custom
event. It enables us to correctly prevent the overlay canceling
by tapping inside the input container, while still having
canceling for tapping anything else outside the date picker.

This commit fixes dependencies and adds iron-overlay-behavior
explicitly with a version requirement. It's necessary since
the date picker implementation currently relies on the original
event presence in the detail of `'iron-overlay-canceled'` event.