Fix for file remove animation in Safari
Issue: in Safari, when removing the file from the list, the animation had a jump in the beginning. Happened in both desktop and mobile Safari.

This commit fixes the jump in animation.

In addition, the file item padding top was slightly reduced to make the total height of the file match 4px grid declared by material design guidelines.