Fix flacky taps scroll tests by listening to scroll event
The scroll event could be delayed for some time after the scrolling
was triggered. This delay matters, for example, when using full Shadow
DOM polyfill, and could make the tests flacky.

This changes the tests to depend on the scroll event rather then
relying on quick enough scroll event dispatch after triggering.

Fixes: vaadin/vaadin-date-picker#334