First iteration of the theme for vaadin-tabs
The theme doesn’t work properly in browsers that rely on ShadyCSS (IE11
and Edge14). This will be fixed once
https://github.com/vaadin/vaadin-tabs/issues/58 is fixed.

The documentation/demo page is still very rough and
Browsers without native shadow DOM have trouble with the
`nice-demo-snippet` used in the examples: the first tab is not selected
automatically, and keyboard navigation doesn’t work. The element works
as expected outside the demo snippet element.

Fixes: vaadin/vaadin-lumo-styles#19