Minor Lumo update [skip ci]
Use min-height instead of a fixed height, so that taller vaadin-items
will grow vaadin-dropdown-menu-text-field. Set min-height on the value
part instead of the input-field, to avoid IE11 flexbox issue.

Move readonly border to a pseudo element to avoid adding a border to
the input-field, so that vaadin-dropdown-menu-text-field height matches
vaadin-text-field height by default (as vaadin-item is as high as the
input-field by default).