Redesign Lumo variants
Base the size of the handle in sizing properties instead of space properties.

Redesigned “small” variant – now just 1px divider and the handle is only visible on hover. The handle is now always the same size for normal and small, only the splitter part changes.

Rename “subtle” to “minimal” (align with other element variant names). The minimal variant doesn’t have any visible splitter, and the handle is only visible on hover.

Update demos to reflect the changes.