Update default Lumo theme
Don’t use the `lumo-overlay` mixin, as notifications are more of a special case, and not that much like the other overlay elements.

Use a dark background color for both light and dark palette. This makes the notications more prominent for the light theme, and easier to notice by the end user.

Increase content padding.