Fix iOS 9 support
Move the mask image to the content part instead of a pseudo element on top of it. There is no added benefit of using the pseudo element really.

Don’t use a custom property for the mask-image (not supported on iOS9). This allows us to remove the Edge workaround as well.