Refactor overlay styles to use lumo-menu-overlay
Remove copy-paste code and use the `lumo-menu-overlay` mixin directly.

Use CSS media queries instead of the `fullscreen` property for the overlay behavior adjustments (consistent with the menu-overlay styles).

A couple of other small style fixes related to sizing and touch device styles.

Use `--vaadin-infinite-scroller-buffer-offset` instead of a transform to shift the first month in the scroller closer to the middle of the overlay.