Generate icons.html and font-icons.html from SVG source files
Add a Gulp script that optimizes the source SVG icons and then generates the iron-iconset-svg and font icon version from those.

The unicode values for the font icons change with this update, but those are not part of the public API, as users should rely on the custom property names instead (which do not change).

Some icons are visually adjusted as well, to make them more coherent.

The individual SVG versions are now also included, if someone wants to use them separately.

The demo page now also allows you to view both the iron-iconset-svg and font icon versions.