Progress bar update
Indeterminate mode is still correct – it’s missing the secondary bar animation. Could try to use a pseudo-element for that.

Had to use a pseudo element for the track so that we can adjust the opacity. The new specs say it should be 6% opacity, but that seems incorrect when looking at the screenshots. The current 16% is chosen visually, which looks close enough.