Postpone overlay stamping to first opening vaadin/vaadin-date-picker#621
This is a performance improvement.

When attaching a lot of date pickers at once, the initial render can take a lot of time. Most of the time is spent on executing logic related to the overlay. This functionality is not needed until the overlay is opened, and it is wasted effort for any date picker on the page that the user doesn't open even once.

As a trade-off the first opening takes a bit longer.