fix: prevent overlay opening on empty input event
The overlay should not be opened when clicking the clear-button. This
functionality broke with a failing test case after text-field 2.4
update, which fires an input-event on clear-button click.

Since we don't know if the input-event was fired from a clear-button
click or keyboard-input, the opening is prevented every time when the
input value is empty. As a tradeoff, this causes a rare issue: When the
input is focused, overlay is not opened and the input is cleared with a
single keypress, the overlay is not opened.