fix: check added checkbox if included in value
In the asynchronous FlattenedNodesObserver, the checkboxes should
be updated to be checked if included in the checkbox-group value.
Previously, this caused an invalid state where the checkbox's
checked-state and the checkbox-group's value were out-of-sync (also when
value was set after adding the checkbox, because of asynchronisity).

fix https://github.com/vaadin/vaadin-checkbox-flow/issues/89