fix: blank page and JS errors in IE11
Fixes: vaadin/vaadin-material-styles#85

In cases where :host selector is used alone for a CSS block in document level styles (e.g. <custom-style>), the block will be broken in at least polyfilled browsers like Edge and IE11 which may break also the following CSS block. In IE11 this seems to also cause some JS errors (seemingly coming from the polyfills) which might stop the whole page from working and only render a blank page. This isn't a problem if the block contains other valid selectors separated with a comma (in that case only the invalid :host selector is automatically removed from document level styles).

This is an issue in cases where we have some style modules which may be used both in shadow DOM and in global scope (via custom-style).

Workaround the issue by adding a valid dummy selector "#host-fix" to style blocks which might be used in this way and only had the ":host" selector.