add a 'transferIssues' script
Intended usage: `node scripts/transferIssues.js` to transfer all open issues from individual web component repos into the web-components monorepo for each of the web components in the ./packages folder of this monorepo.

Labels on the transferred issues are transferred as well. The transferred issues remain in the same pipelines (columns) on the DS team ZH boards.

If a label was missing in the web-components repo, it will be created (with the same color as in the original repo).