transfer issues, labels and pipelines
Also add more filtering to have a way to limit the scope of this script, which is handy for testing.

To run this script 'for real' execute `PRODUCTION_RUN=true node scripts/transferIssues.js`. With the filters as they are committed right now this script would transfer a single issue: https://github.com/vaadin/magi-cli/issues/142