chore: add a `transferFlowIssues` script [skip ci] (#1975)
Intended usage: `node scripts/transferFlowIssues.js` to transfer all open issues with the `flow` label from the `web-components` repo into `flow-components`. In order to run with the `vaadin-bot` credentials, run this script via the [Transfer issues to the flow-components monorepo](https://bender.vaadin.com/buildConfiguration/VaadinWebComponents_TransferIssuesToTheFlowComponentsMonorepo/232264?buildTab=log&expandAll=true&focusLine=2909) CI build.

Labels on the transferred issues are transferred as well (except for the `flow` label, which is dropped). The transferred issues remain in the same pipelines (columns) on all ZenHub boards.

If a label was missing in the web-components repo, it will be created (with the same color as in the original repo).