docs: deprecate the `iron-resize` event [skip ci]
The `iron-resize` event is an internal implementation detail, yet it's currently documented as a part of the public API of:
 - vaadin-text-field
 - vaadin-number-field
 - vaadin-email-field
 - vaadin-password-field
 - vaadin-text-area
 - vaadin-split-layout

This commit adda a depracation note for this event in  the API docs. The intention is to drop this even in the future when the implementat will change from IronResizableBehavior to ResizeObserver.

fixes #2167