fix(grid): skip calculation when Grid is hidden
If a resize or any event that triggers `iron-resize` happens as the
Grid is about to hide, it may lead to some calculations to happen
when the Grid is already hidden. That may lead to some misbehavior,
like described on the linked issue.

This fix checks the visibility status of the Grid to skip metrics
calculation in case the Grid is not visible.

fix https://github.com/vaadin/web-components/issues/2127