fix: use isVisible on iron-list to check Grid's visibility
This is a complementary fix of #2244.
The original fix works when Grid is directly set to hidden, but fails
for other cases, eg. when a Grid's parent get hidden instead.

This change uses iron-list`s _isVisible to check Grid's visibility
to determine whether any further calculation should be skipped or not.

Related to #2127