feat: expose editor buttons with slots
This is an internal API to be used by the Flow counterpart.
Expose the editor buttons (save, cancel and delete) with slots in a
similar way as done with form. This is to enable the Flow counterpart
to pass its own buttons instances and attach them to the Crud WC.

Part of https://github.com/vaadin/flow-components/issues/1112

Co-authored-by: Hadi Amiri <hdamr@users.noreply.github.com>