feat: theme property for individual menu items
The vaadin-menu-bar as a whole is themable, and setting the theme attribute on it will propagate it to all menu bar buttons.
However, a user might want to add a theme to only some select buttons, for example to show one as the primary option. While
this can be done by supplying a component, it is not the most convenient approach, and the theme property is also needed
for the Java API. Part of vaadin/flow-components#880.