fix: make header-title be announced by screen readers
Make header-title set `aria-label` instead of using `aria-labelleby` as
it seems that Chrome + VoiceOver is not capable of announce the later.
Make sure that the `aria-label` property set on the dialog is set in
case both `aria-label` and `header-title` are defined.
Change the markup of the resize container from `<div>` to `<section>`.

Fix #3702