refactor!: use aria-describedby for group field components
As both JAWS and NVDA announce aria-describedby correctly,
and only VoiceOver is unable to do so, we'll revert to
using aria-describedby for content that is not semantically
part of the label.

Otherwise we would have to refactor tooltips to also use
aria-labelledby, as well as introduce a hidden element for
invisible aria-labels. Futhermore, concatenating label,
description, validation errors and tooltips into the
label results in extremely verbose announcements.

We see this as choosing the less bad (and less complex)
of two bad options.