fix: use drawer effecive size for content offset
This change brings back the calculation done for the offset
size used as the padding in the content area.
It uses the actual drawer size instead of relying only on the
new CSS variable, since the drawer size can be defined in different
ways (e.g. `vaadin-app-layout::part(drawer) { width: 200px; }`).