fix: avoid clearing filter when pressing clear button on mobile
Disables the clear button logic from `vaadin-combo-box-mixin`, which clears the filter input when pressing the clear button. Instead, only the MSCB clear button logic is executed, which only clears the selection.

The bug actually only occurs on the first touch event, and not on subsequent ones. The root cause is that on the first clear, `selectedItem` is changed from `undefined` to `null`, which runs the respective change logic that also resets the input. Ideally `selectedItem` would be initialized with `null`, so the change would never happen. However, doing so results in a lot of test failures. I guess it would require additional changes to handle the initial Polymer change event then, but if someone has an idea feel free to make suggestions.