fix: update media query to work with larger phones
Update the media query that sets the overlay to show on the bottom of
the page to work with larger phones, such as the iPhone Pro Max.

For such devices, this current media query is not met leading to the
menu overlay being opened at the top-left corner of the screen.