refactor: open date picker overlay on element focus in fullscreen
This change overrides the `focus` method to change how focus is handled
when the element is in fullscreen mode and it's focused. This is needed
to avoid the element being immediatelly removed by the Grid Pro when the
user is in a device with small viewport. The element receives focus, but
the `_onFocus` listener has a similar detection which removes the focus
from the input element, which causes the component to be removed from
the cell when used as a custom editor.

Part of #9158