fix: listen to `opened-changed` events in custom editors
This change adds a `opened-changed` listener for custom editors and
checks that either the target element or its `inputElement` is focused.
If not, then calls the routine to stop the edit in the cell.

This is done for the case a `vaadin-date-pikcer` is used as a editor and
the user is interacting with the cell containing it in a device with a
small viewport like a phone. In this case, when the picker overlay is
closed, the focus is not restored to the input, which should be a right
time to remove the editor.

Part of #9158