feat: add `chart-end-resize` event
Add a new event that fires when the callback to
`Highcharts.Chart#event:endResize` is called. It is being added now to
make tests that check for the size of a chart after a resize is
performed to be more reliable and to work better when the Highcharts
version is upgraded.

Part of #9908