fix: restrict custom data label colors
To ensure a good contrast between the background color of a series and
its data label color, a custom color is calculated and set to the data
labels. That works well for some types of charts where the labels are
placed within an area colored by the series color, like for
Organization, Gantt, and Treemap. For other types, however, it can lead
to unexpected results, when the data labels are placed in an area
covered by the chart's background, when `useHTML` is used.

This change restricts the setting of these custom colors to the types
mentioned above, where they are expected to work well.

Fixes #9988