fix: force size calculation on reattach
When moving a `<vaadin-chart>` instance to a different container with
different sizes, it might end up with an unexpected size, especially in
cases where it moves from a container without a defined height to a new
one with a defined height and back. It is expected that it would restore
to the original height of the first parent, but it might keep the height
of the second parent.

This change forces a recalculation of size by unsetting the `wrapper`
height, then calling the `setSize` method in the configuration and
resetting the height back to its original value.

That will cause the chart to go back to its default chart size.

Fixes #9886